+++ transport/errors.go
@@ -2,8 +2,8 @@ package transport
 
 import "errors"
 
-var errTimeout = errors.New("Transport: timeout")
+var ErrTimeout = errors.New("Transport: timeout")
 
 func IsTimeout(err error) bool {
-	return err == errTimeout
+	return err == ErrTimeout
 }
